Moving leaderelectionconfig to k8s.io/component-base
