+++ cmd/controller-manager/app/options/BUILD
@@ -12,12 +12,12 @@ go_library(
     importpath = "k8s.io/kubernetes/cmd/controller-manager/app/options",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/client/leaderelectionconfig:go_default_library",
         "//pkg/controller/apis/config:go_default_library",
         "//pkg/controller/service/config:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
         "//staging/src/k8s.io/component-base/config:go_default_library",
+        "//staging/src/k8s.io/component-base/config/options:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
     ],
 )
